Move OpPayloadBuilderCtx to dedicated crate
