+++ crates/op-rbuilder/src/primitives/mod.rs
@@ -1,6 +1,7 @@
 //! This module contains types from the reth that weren't heavily modified
 mod execution;
 mod helpers;
+mod payload_builder_ctx;
 mod payload_builder_service;
 
 pub use payload_builder_service::PayloadBuilderService;
@@ -8,3 +9,5 @@ pub use payload_builder_service::PayloadBuilderService;
 pub use helpers::{estimate_gas_for_builder_tx, signed_builder_tx};
 
 pub use execution::{ExecutedPayload, ExecutionInfo};
+
+pub use payload_builder_ctx::OpPayloadBuilderCtx;
